Online & SavePatternSystem
Implemented savepattern system in PVP mode. Changed .net 2 to .net4xxx or whatever that was, so savepatternsystem's asset could be built. Added exit and save buttons to the toolbar prefab. Fixed WindowScript so when a player decides to close the aplication via toolbar's close button, connection in PVP scene could be closed correctly as well